Solver optimizations
