+++ utbot-framework/src/main/kotlin/org/utbot/engine/selectors/nurs/InheritorsSelector.kt
@@ -1,7 +1,7 @@
 package org.utbot.engine.selectors.nurs
 
 import org.utbot.engine.ExecutionState
-import org.utbot.engine.TypeRegistry
+import org.utbot.engine.types.TypeRegistry
 import org.utbot.engine.selectors.strategies.DistanceStatistics
 import org.utbot.engine.selectors.strategies.StoppingStrategy
 import kotlin.math.pow
