+++ utbot-framework/src/main/kotlin/org/utbot/engine/types/TypeResolver.kt
@@ -1,11 +1,35 @@
-package org.utbot.engine
+package org.utbot.engine.types
 
 import org.utbot.common.WorkaroundReason
 import org.utbot.common.workaround
+import org.utbot.engine.ArrayValue
+import org.utbot.engine.ChunkId
+import org.utbot.engine.Hierarchy
+import org.utbot.engine.MemoryChunkDescriptor
+import org.utbot.engine.ObjectValue
+import org.utbot.engine.ReferenceValue
+import org.utbot.engine.TypeStorage
+import org.utbot.engine.appropriateClasses
+import org.utbot.engine.baseType
+import org.utbot.engine.findMockAnnotationOrNull
+import org.utbot.engine.isAppropriate
+import org.utbot.engine.isArtificialEntity
+import org.utbot.engine.isInappropriate
+import org.utbot.engine.isJavaLangObject
+import org.utbot.engine.isLambda
+import org.utbot.engine.isLocal
+import org.utbot.engine.isOverridden
+import org.utbot.engine.isUtMock
+import org.utbot.engine.makeArrayType
+import org.utbot.engine.nullObjectAddr
+import org.utbot.engine.numDimensions
 import org.utbot.engine.pc.UtAddrExpression
 import org.utbot.engine.pc.UtBoolExpression
 import org.utbot.engine.pc.mkAnd
 import org.utbot.engine.pc.mkEq
+import org.utbot.engine.rawType
+import org.utbot.engine.wrapper
+import org.utbot.engine.wrapperToClass
 import org.utbot.framework.plugin.api.id
 import org.utbot.framework.plugin.api.util.id
 import org.utbot.framework.util.UTBOT_FRAMEWORK_API_VISIBLE_PACKAGE
